Backend/requirements.txt
cryptography
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-PyMongo
Flask-Migrate
Flask-Migrate
Flask-Cors
Flask-mongoengine
python-dotenv
pymysql
requests
requests-oauthlib
apscheduler
scikit-learn
xgboost
numpy
pandas
werkzeug
google-auth
google-api-python-client